+++ projects/egeo/src/lib/st-menu/st-menu.model.ts
@@ -1,4 +1,4 @@
-/**
+/*
  * © 2017 Stratio Big Data Inc., Sucursal en España.
  *
  * This software is licensed under the Apache License, Version 2.0.
@@ -8,8 +8,25 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
-@import 'constants/index';
 
-:host {
-   width: 100%;
+export interface StMenuModel<T> {
+  name: string;
+  value: T;
+  icon?: string;
+  subMenus?: Array<StMenuModel<T>>;
+  separator?: boolean;
+  status?: StMenuStatus;
+  disabled?: boolean;
+}
+
+
+export enum StMenuStatus {
+   success = 'success',
+   warning = 'warning',
+   critical = 'critical'
+}
+
+export enum StMenuPosition {
+   left = 'left',
+   right = 'right'
 }